requirements.txt
blinker==1.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
future==0.18.2
idna==2.8
oauthlib==3.0.1
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==19.0.0
PySocks==1.7.1
python-twitter==3.5
requests==2.22.0
requests-oauthlib==1.2.0
six==1.13.0
urllib3==1.25.7
